Add aot_autograd.fx_utils
See docblock for details.  The API here has been validated by use
in autoparallel but I'm always open to suggestions for tweaks.  One
particular choice I made is to make most of the functions return dicts
by default; this isn't strictly necessary for inputs but it is very
convenient for outputs as the output desc lives on the output node,
not the argument that feeds into the node.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
ghstack-source-id: a0962ca397d515efb64f896f747a4c919bf4fd11
Pull-Request: https://github.com/pytorch/pytorch/pull/159005